requirements.txt
praisonaiagents[knowledge]
python-dotenv
ollama==0.1.24
chromadb
streamlit
